Update test_fused_layer_norm.py (#1485)
* first commit

* remove unnecessary imports, move prep layers into functions

* added minor fix for type conversion in manual_rms_norm

+++ apex/normalization/fused_layer_norm.py
@@ -24,7 +24,7 @@ def manual_rms_norm(input, normalized_shape, weight, eps):
 
     # convert into half-precision if necessary
     if weight.dtype in [torch.float16, torch.bfloat16]:
-        input = input.to(self.weight.dtype)
+        input = input.to(weight.dtype)
 
     return weight * input
